requirements-startup.txt
Pillow>=5.3.0
PyGithub
aiofiles
aiohttp
aria2p
async_generator
beautifulsoup4
bs4
bwb==3.0.0
cfscrape
coffeehouse
cowpy
cryptg
dnspython
emoji
gTTS-token>=1.1.3
gTTS>=2.0.1
geopy
gitpython
google-api-python-client==1.8.0
google-auth-oauthlib
google_images_download>=2.7.1
googletrans>=2.4.0
gsearch
hachoir
heroku3
httplib2
humanize
lxml
oauth2client
psycopg2
psycopg2-binary
PyLyrics
pySmartDL
pybase64>=0.4.0
pyfiglet
pylast
pymongo
python-barcode
python-dotenv
python-magic
pytube
pytz
qrcode
regex
requests>=2.18.4
search-engine-parser>=0.4.2
selenium
speedtest-cli>=2.0.2
sqlalchemy>=1.2
telegraph
urbandict>=0.5
wikipedia>=1.4.0
youtube-dl
covid
psutil
